WIP resolve form submission errors
Original error was displayed when trying to create new project - state of isPreAllocated was always undefned. This issue has been resolved in this commit.

One student could be allocated to many projects. Work started to resolve this issue. Current state of commit leads to student not being deallocated from a project correctly - possible solution is update to edit procedure condition when a different student is allocated to a project ("correct" behaviour can be observed when changing from student defined project to a normal project and vice versa).

Co-authored-by: Petros <pkitazos@users.noreply.github.com>