All flags & tags can be removed from an instance and/or project
Previous changes allowed an admin to remove flags and tags from an instance which could result in a project having no flags or tags.
If a project with no flags or tags was edited, an error would be raised as form validation required at least 1 flag and tag.

Removed min values for flags and tags in the project form schema and removed rules flags form validation so that flags array length can be 0.

Project id page will now only show flags and/or tags in project details (on side of page) if the project has flags and/or tags.

If an instance does not have any flags, then the option to add flags has been removed from create and edit project form.