Fix student rank error decreasing after manual adjustment
When clicking on "Submit Changes" on manual changes page, updateAllocation procedure in matching is called. updateAllocation uses getStudentRank to update the student ranking. getStudentRank uses getProjectRank to update the student ranking which returns the index of the allocated project in the students preferences array.

This means that if a student was allocated their first choice, the index would be 0, and so the student ranking was updated to 0.

Added 1 to returned index.