Allow admins to submit projects.
This has a few changes, and needs further testing.

The "UserInInstance.Role" field should just be either STUDENT or
SUPERVISOR. Instead of storing ADMIN there, use "AdminInSpace" table.

The api.user.adminLevel / adminLevelInGroup / adminLevelInSubGroup
procedures are now just protectedProcedure, not adminProcedure, so that
regular users can check if they are admins. This is needed because:

If someone has the SUPERVISOR role, but is an admin of some sort, they
are shown the admin panel and not the supervisor panel for instances.
This does mean that they don't see their task list or similar.

For these admins who are supervisors, we just manually add the "Create
project" / "My Projects" buttons for them depending on the stage.

Lastly, the lowest AdminLevel is now AdminLevel.NONE.